pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tpc-reporter"
version = "0.1.0"
description = "Generate track reports from TPC conference data"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "TPC Workshop Reporter Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pyyaml>=6.0",
    "openai>=1.0.0",
    "click>=8.0",
]

[project.optional-dependencies]
google = [
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
]
scraper = [
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
all = [
    "tpc-reporter[google,scraper,dev]",
]

[project.scripts]
tpc-reporter = "tpc_reporter.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tpc_reporter*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.ruff_cache
    | __pycache__
    | \.pytest_cache
    | originals
    | archive
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = [
    "E501",  # Line too long (handled by black)
    "E402",  # Module level import not at top (conflicts with __future__ + docstring)
]

[tool.ruff.lint.isort]
known-first-party = ["tpc_reporter"]
